refactor(org): reorganise folder structure
