+++ src/components/disclaimer/disclaimer.css
