+++ src/components/routes/routes.jsx
@@ -1,6 +1,6 @@
 import React, {Component} from "react";
-import logo_horizontal from "../assets/cgw_logo+text+horizontal.png";
-import Home from "../home";
+import logo_horizontal from "../../assets/cgw_logo+text+horizontal.png";
+import Home from "../../home";
 import "./routes.css";
 
 let lastScrollTop = 0;
